requirements.txt
jupyter
numpy
scipy
matplotlib
colour-science
colour-checker-detection
scikit-image
